# backend/pkg/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  ErrorResponse:
    properties:
      code:
        example: Internal
        type: string
      error:
        example: original server error message
        type: string
      msg:
        example: internal server error
        type: string
      status:
        example: error
        type: string
    type: object
  SuccessResponse:
    properties:
      data:
        type: object
      status:
        example: success
        type: string
    type: object
...
  services.agentlogs:
    properties:
      agentlogs:
        items:
          $ref: '#/definitions/models.Agentlog'
        type: array
      total:
        type: integer
    type: object
  services.anonymizeTextRequest:
    properties:
      text:
        type: string
    required:
    - text
    type: object
  services.anonymizeTextResponse:
    properties:
      text:
        type: string
    type: object
  services.assistantlogs:
    properties:
      assistantlogs:
        items:
          $ref: '#/definitions/models.Assistantlog'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact:
    email: team@pentagi.com
    name: PentAGI Development Team
    url: https://pentagi.com
  description: Swagger API for Penetration Testing Advanced General Intelligence PentAGI.
  license:
    name: MIT
    url: https://opensource.org/license/mit
  termsOfService: http://swagger.io/terms/
  title: PentAGI Swagger API
  version: "1.0"
paths:
  /agentlogs/:
    get:
      parameters:
      - collectionFormat: multi
        description: |-
          Filtering result on server e.g. {"value":[...],"field":"...","operator":"..."}
            field is the unique identifier of the table column, different for each endpoint
            value should be integer or string or array type, "value":123 or "value":"string" or "value":[123,456]
            operator value should be one of <,<=,>=,>,=,!=,like,not like,in
            default operator value is 'like' or '=' if field is 'id' or '*_id' or '*_at'
        in: query
        items:
          type: string
        name: filters[]
        type: array
      - description: Field to group results by
        in: query
        name: group
        type: string
      - default: 1
        description: Number of page (since 1)
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - default: 5
        description: Amount items per page (min -1, max 1000, -1 means unlimited)
        in: query
        maximum: 1000
        minimum: -1
        name: pageSize
        type: integer
      - collectionFormat: multi
        description: |-
          Sorting result on server e.g. {"prop":"...","order":"..."}
            field order is "ascending" or "descending" value
            order is required if prop is not empty
        in: query
        items:
          type: string
        name: sort[]
        type: array
      - default: init
        description: Type of request
        enum:
        - sort
        - filter
        - init
        - page
        - size
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: agentlogs list received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.agentlogs'
              type: object
        "400":
          description: invalid query request data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: getting agentlogs not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on getting agentlogs
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve agentlogs list
      tags:
      - Agentlogs
  /anonymize/text:
    post:
      consumes:
      - application/json
      description: Replace sensitive data patterns (credentials, keys, PII, etc.)
        in the provided text with safe placeholders.
      parameters:
      - description: Text to anonymize
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/services.anonymizeTextRequest'
      produces:
      - application/json
      responses:
        "200":
          description: anonymized text returned successfully
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.anonymizeTextResponse'
              type: object
        "400":
          description: invalid request body
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: anonymize.call permission required
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: anonymizer is not configured
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Anonymize text
      tags:
      - Anonymize
  /assistantlogs/:
    get:
      parameters:
      - collectionFormat: multi
        description: |-
          Filtering result on server e.g. {"value":[...],"field":"...","operator":"..."}
            field is the unique identifier of the table column, different for each endpoint
            value should be integer or string or array type, "value":123 or "value":"string" or "value":[123,456]
            operator value should be one of <,<=,>=,>,=,!=,like,not like,in
            default operator value is 'like' or '=' if field is 'id' or '*_id' or '*_at'
        in: query
        items:
          type: string
        name: filters[]
        type: array
      - description: Field to group results by
        in: query
        name: group
        type: string
      - default: 1
        description: Number of page (since 1)
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - default: 5
        description: Amount items per page (min -1, max 1000, -1 means unlimited)
        in: query
        maximum: 1000
        minimum: -1
        name: pageSize
        type: integer
      - collectionFormat: multi
        description: |-
          Sorting result on server e.g. {"prop":"...","order":"..."}
            field order is "ascending" or "descending" value
            order is required if prop is not empty
        in: query
        items:
          type: string
        name: sort[]
        type: array
      - default: init
        description: Type of request
        enum:
        - sort
        - filter
        - init
        - page
        - size
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: assistantlogs list received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.assistantlogs'
              type: object
        "400":
          description: invalid query request data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: getting assistantlogs not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on getting assistantlogs
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve assistantlogs list
      tags:
      - Assistantlogs
  /auth/authorize:
    get:
      parameters:
      - default: /
        description: URI to redirect user there after login
        in: query
        name: return_uri
        type: string
      - default: google
        description: OAuth provider name (google, github, etc.)
        in: query
        name: provider
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: redirect to SSO login page
        "400":
          description: invalid autorizarion query
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: authorize not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on autorizarion
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Login user into OAuth2 external system via HTTP redirect
      tags:
      - Public
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login form JSON data
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.Login'
      produces:
      - application/json
      responses:
        "200":
          description: login successful
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: invalid login data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: invalid login or password
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: login not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on login
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Login user into system
      tags:
      - Public
  /auth/login-callback:
    get:
      consumes:
      - application/json
      parameters:
      - description: Auth code from OAuth provider to exchange token
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "303":
          description: redirect to registered return_uri path in the state
        "400":
          description: invalid login data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: invalid login or password
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: login not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on login
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Login user from external OAuth application
      tags:
      - Public
    post:
      consumes:
      - application/json
      parameters:
      - description: Auth form JSON data
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.AuthCallback'
      produces:
      - application/json
      responses:
        "303":
          description: redirect to registered return_uri path in the state
        "400":
          description: invalid login data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: invalid login or password
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: login not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on login
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Login user from external OAuth application
      tags:
      - Public
  /auth/logout:
    get:
      parameters:
      - default: /
        description: URI to redirect user there after logout
        in: query
        name: return_uri
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: redirect to input return_uri path
      summary: Logout current user via HTTP redirect
      tags:
      - Public
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"